Check team permissions at boot
